ml_sidecar/requirements.txt
pandas>=2.2,<3
numpy>=2.0,<3
scikit-learn>=1.5,<2
xgboost>=2.1,<3
supabase>=2.15,<3
python-dotenv>=1.0,<2
holidays>=0.70,<1
